refactor(liquidity monitor): PLA-25 Fixed review comments
