+++ packages/liquidity-monitor-backend/src/provider-monitoring/provider-monitoring.module.ts
@@ -2,6 +2,7 @@ import { Module } from "@nestjs/common";
 import { ProviderMonitoringService } from "./provider-monitoring.service";
 import { TelegrafModule } from "@boarteam/boar-pack-users-backend";
 import { ProviderMonitoringController } from "./provider-monitoring.controller";
+import { SettingsModule } from "@boarteam/boar-pack-users-backend";
 
 @Module({})
 export class ProviderMonitoringModule {
@@ -12,6 +13,10 @@ export class ProviderMonitoringModule {
       module: ProviderMonitoringModule,
       imports: [
         TelegrafModule.register({
+          withControllers: false,
+          dataSourceName: config.dataSourceName,
+        }),
+        SettingsModule.register({
           withControllers: false,
           dataSourceName: config.dataSourceName,
         }),
