+++ packages/users-backend/src/settings/settings.constants.ts
@@ -1,9 +1,8 @@
-export const Telegram = {
-  BotToken: 'telegram.bot_token',
-  ChatId: 'telegram.chat_id',
-  Enabled: 'telegram.enabled',
-} as const;
-
 export enum Notifications {
   QuotesByProviderStatus = 'notifications.quotes_by_provider_status',
+}
+
+export enum SettingsValues {
+  NO = 'no',
+  YES = 'yes',
 }
