+++ packages/users-backend/src/telegraf/index.ts
@@ -1,2 +1,3 @@
 export * from './telegraf.module';
 export * from './telegraf.service';
+export * from './telegraf.constants';
