+++ packages/users-backend/src/telegraf/telegraf.constants.ts
@@ -1,0 +1,5 @@
+export const Telegram = {
+  BotToken: 'telegram.bot_token',
+  ChatId: 'telegram.chat_id',
+  Enabled: 'telegram.enabled',
+} as const;
